🚚(frontend) rename CourseProductContext
With CERTIFICATE product evolution, this context must
now handle course and enrollment.
The new name ProductRelationContext is a better fit.
