+++ src/frontend/js/components/PurchaseButton/index.spec.tsx
@@ -9,7 +9,7 @@ import {
 } from 'utils/test/factories/richie';
 import { CourseRunFactory, ProductFactory } from 'utils/test/factories/joanie';
 import { SessionProvider } from 'contexts/SessionContext';
-import { CourseProductProvider } from 'contexts/CourseProductContext';
+import { ProductRelationProvider } from 'contexts/ProductRelationContext';
 import { createTestQueryClient } from 'utils/test/createTestQueryClient';
 import { ProductType } from 'types/Joanie';
 import { Priority } from 'types';
@@ -36,13 +36,13 @@ describe('PurchaseButton', () => {
   });
 
   const Wrapper = ({ client, children }: React.PropsWithChildren<{ client: QueryClient }>) => (
-    <CourseProductProvider courseCode="00000" productId="00000">
+    <ProductRelationProvider courseCode="00000" productId="00000">
       <IntlProvider locale="en">
         <QueryClientProvider client={client}>
           <SessionProvider>{children}</SessionProvider>
         </QueryClientProvider>
       </IntlProvider>
-    </CourseProductProvider>
+    </ProductRelationProvider>
   );
 
   it('shows a login button if user is not authenticated', async () => {
