+++ src/frontend/js/contexts/ProductRelationContext/index.spec.tsx
@@ -1,20 +1,20 @@
 import { renderHook } from '@testing-library/react';
 import { noop } from 'utils';
-import { CourseProductProvider, useCourseProduct } from '.';
+import { ProductRelationProvider, useCourseProduct } from '.';
 
 describe('useCourseProduct', () => {
   afterEach(() => {
     jest.restoreAllMocks();
   });
 
-  it('returns the course code stored within CourseProductProvider', () => {
+  it('returns the course code stored within ProductRelationProvider', () => {
     // Refactor of the wrapper parameters is caused by the behavior describe here
     // https://stackoverflow.com/questions/72346908/react-testing-library-merged-with-testing-library-react-hooks-wrapper-inconsis
     const { result } = renderHook(useCourseProduct, {
       wrapper: ({ children }) => (
-        <CourseProductProvider courseCode="00013" productId="dce9acb4-df59-4856-a5a8-1b732bc9b565">
+        <ProductRelationProvider courseCode="00013" productId="dce9acb4-df59-4856-a5a8-1b732bc9b565">
           {children}
-        </CourseProductProvider>
+        </ProductRelationProvider>
       ),
     });
     expect(result.current).toStrictEqual({
@@ -24,12 +24,12 @@ describe('useCourseProduct', () => {
     });
   });
 
-  it('throws an error if it is not used within a CourseProductProvider', () => {
+  it('throws an error if it is not used within a ProductRelationProvider', () => {
     jest.spyOn(console, 'error').mockImplementation(noop);
     expect(() => {
       renderHook(useCourseProduct);
     }).toThrow(
-      'useCourseProduct must be used within a component wrapped by a <CourseProductProvider />.',
+      'useCourseProduct must be used within a component wrapped by a <ProductRelationProvider />.',
     );
   });
 });
